+++ tests/TestCase/Routing/Middleware/ThumbnailMiddlewareTest.php
@@ -89,7 +89,7 @@ class ThumbnailMiddlewareTest extends TestCase
         $this->assertResponseCode(304);
 
         //Deletes the last thumbnail file. Now the `Last-Modified` header is different
-        safe_unlink($thumb);
+        @unlink($thumb);
         sleep(1);
         (new ThumbCreator($file))->resize(200)->save();
         $this->get($url);