+++ tests/TestCase/Utility/ThumbCreatorSaveTest.php
@@ -123,7 +123,7 @@ class ThumbCreatorSaveTest extends TestCase
 
         //Deletes the thumbnail and wait 1 second, then tries to create again
         //  the same thumbnail. Now the creation time is different
-        safe_unlink($thumb);
+        @unlink($thumb);
         sleep(1);
         $newTime = filemtime($this->getThumbCreatorInstance()->resize(200)->save());
         $this->assertNotEquals($time, $newTime);