+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ## 1.7 branch
 # 1.7.1
 * renamed commands. Now they are `thumber.clear` and `thumber.clear_all`;
-* updated for CakePHP 3.7.1 and `php-tools` 1.1.19.
+* updated for CakePHP 3.7.1 and `php-tools` 1.1.12.
 
 # 1.7.0
 * `ThumberShell` has been replaced with console commands. Every method of the